requirements.txt
attrs==22.1.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Fiona==1.8.22
GeoAlchemy2==0.12.5
geopandas==0.10.2
greenlet==2.0.1
importlib-metadata==5.2.0
joblib==1.2.0
munch==2.5.0
numpy==1.21.6
packaging==22.0
pandas==1.3.5
pyproj==3.2.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
shapely==2.0.0
six==1.16.0
SQLAlchemy==1.4.45
typing_extensions==4.4.0
zipp==3.11.0
